#24778 adding mergeLatest graph stage
