+++ akka-stream/src/main/scala/akka/stream/scaladsl/Graph.scala
@@ -20,7 +20,6 @@ import scala.annotation.tailrec
 import scala.annotation.unchecked.uncheckedVariance
 import scala.collection.{ immutable, mutable }
 import scala.concurrent.Promise
-import scala.reflect.ClassTag
 import scala.util.control.{ NoStackTrace, NonFatal }
 
 /**
@@ -484,75 +483,6 @@ final class Interleave[T](val inputPorts: Int, val segmentSize: Int, val eagerCl
   }
 
   override def toString = "Interleave"
-}
-
-object MergeLatest {
-  /**
-   * Create a new `MergeLatest` with the specified number of input ports.
-   *
-   * @param inputPorts number of input ports
-   * @param eagerComplete if true, the merge latest will complete as soon as one of its inputs completes.
-   */
-  def apply[T: ClassTag](inputPorts: Int, eagerComplete: Boolean = false): MergeLatest[T] = new MergeLatest[T](inputPorts, eagerComplete)
-
-}
-
-/**
- * MergeLatest joins elements from N input streams into stream of lists of size N.
- * i-th element in list is the latest emitted element from i-th input stream.
- * MergeLatest emits list for each element emitted from some input stream,
- * but only after each stream emitted at least one element
- *
- * '''Emits when''' element is available from some input and each input emits at least one element from stream start
- *
- * '''Backpressures when''' downstream backpressures
- *
- * '''Completes when''' all upstreams complete (eagerClose=false) or one upstream completes (eagerClose=true)
- *
- * '''Cancels when''' downstream cancels
- *
- */
-final class MergeLatest[T: ClassTag](val inputPorts: Int, val eagerClose: Boolean) extends GraphStage[UniformFanInShape[T, List[T]]] {
-  require(inputPorts >= 1, "input ports must be >= 1")
-
-  val in: immutable.IndexedSeq[Inlet[T]] = Vector.tabulate(inputPorts)(i ⇒ Inlet[T]("MergeLatest.in" + i))
-  val out: Outlet[List[T]] = Outlet[List[T]]("MergeLatest.out")
-  override val shape: UniformFanInShape[T, List[T]] = UniformFanInShape(out, in: _*)
-
-  override def createLogic(inheritedAttributes: Attributes): GraphStageLogic = new GraphStageLogic(shape) with OutHandler {
-    private val activeStreams: mutable.HashSet[Int] = new mutable.HashSet[Int]()
-    private var runningUpstreams: Int = inputPorts
-    private def upstreamsClosed: Boolean = runningUpstreams == 0
-    private def allMessagesReady: Boolean = activeStreams.size == inputPorts
-    private val messages: Array[T] = new Array[T](inputPorts)
-
-    in.zipWithIndex.foreach {
-      case (input, index) ⇒
-        setHandler(input, new InHandler {
-          override def onPush(): Unit = {
-            messages.update(index, grab(input))
-            activeStreams.add(index)
-            if (allMessagesReady) emit(out, messages.toList)
-            tryPull(input)
-          }
-
-          override def onUpstreamFinish(): Unit = {
-            if (!eagerClose) {
-              runningUpstreams -= 1
-              if (upstreamsClosed) completeStage()
-            } else completeStage()
-          }
-        })
-    }
-
-    def onPull(): Unit =
-      for (index ← 0 until inputPorts)
-        if (!hasBeenPulled(in(index))) tryPull(in(index))
-
-    setHandler(out, this)
-  }
-
-  override def toString = "MergeLatest"
 }
 
 /**
